Revert "Renamed CASMonitorHttpClient to the more specific DataHubMonitorHttpClient. Moved the class to a datahub specific location."
This reverts commit 0a07469c21bbe58ac618b2a1dae71a4dcaa7a8c1.
