+++ engine/core/src/main/java/org/datacleaner/metamodel/datahub/DataHubConnection.java
@@ -28,6 +28,7 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.client.HttpClients;
 import org.datacleaner.util.SecurityUtils;
+import org.datacleaner.util.http.CASMonitorHttpClient;
 import org.datacleaner.util.http.HttpBasicMonitorHttpClient;
 import org.datacleaner.util.http.MonitorHttpClient;
 
@@ -74,7 +75,7 @@ public class DataHubConnection {
         final CloseableHttpClient httpClient = clientBuilder.build();
 
         if (CAS.equals(_securityMode)) {
-            return new DataHubMonitorHttpClient(httpClient, getCasServerUrl(), _username, _password, contextUrl);
+            return new CASMonitorHttpClient(httpClient, getCasServerUrl(), _username, _password, contextUrl);
         } else {
             return new HttpBasicMonitorHttpClient(httpClient, getHostname(), getPort(), _username, _password);
         }
@@ -93,7 +94,7 @@ public class DataHubConnection {
         final CloseableHttpClient httpClient = clientBuilder.build();
 
         if (CAS.equals(_securityMode)) {
-            return new DataHubMonitorHttpClient(httpClient, getCasServerUrl(), _username, _password, contextUrl);
+            return new CASMonitorHttpClient(httpClient, getCasServerUrl(), _username, _password, contextUrl);
         } else {
             return new HttpBasicMonitorHttpClient(httpClient, getHostname(), getPort(), _username, _password);
         }
