+++ engine/core/src/main/java/org/datacleaner/util/http/CASMonitorHttpClient.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.metamodel.datahub;
+package org.datacleaner.util.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -51,7 +51,6 @@ import org.apache.http.protocol.HttpContext;
 import org.apache.http.util.EntityUtils;
 import org.apache.metamodel.util.FileHelper;
 import org.apache.metamodel.util.LazyRef;
-import org.datacleaner.util.http.MonitorHttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -62,9 +61,9 @@ import org.slf4j.LoggerFactory;
  * This client requires that CAS is installed with the RESTful API, which is
  * described in detail here: https://wiki.jasig.org/display/CASUM/RESTful+API
  */
-public class DataHubMonitorHttpClient implements MonitorHttpClient {
+public class CASMonitorHttpClient implements MonitorHttpClient {
 
-    private static final Logger logger = LoggerFactory.getLogger(DataHubMonitorHttpClient.class);
+    private static final Logger logger = LoggerFactory.getLogger(CASMonitorHttpClient.class);
 
     private static final String CDI_TICKET_HEADER = "CDI-ticket";
     private static final String CDI_SERVICE_URL_HEADER = "CDI-serviceUrl";
@@ -81,7 +80,7 @@ public class DataHubMonitorHttpClient implements MonitorHttpClient {
     private String _requestedService;
     private String _casRestServiceUrl;
 
-    public DataHubMonitorHttpClient(CloseableHttpClient client, String casServerUrl, String username, String password,
+    public CASMonitorHttpClient(CloseableHttpClient client, String casServerUrl, String username, String password,
             String monitorBaseUrl) {
         _httpClient = client;
         _casServerUrl = casServerUrl;
